avoid OpenPGP.js when building
GnuPG should not depend on OpenPGP.js.

This is an attempt to strip out whatever dependency it had and make it
work with gpg alone.

There is no need for getStrippedKey at all -- it's never invoked in
the codebase outside of getMinimalPubKey, and it should not be
relevant there, since we are already filtering appropriately.

Since we explicitly depend on gnupg 2.2.8-2 or greater, we are fine
with these expectations.

We also don't want to try to test openpgp.js at all.