use GnuPG instead of openpgp.js for symmetric encryption/decryption
This enables all the test suite to pass, and should enable proper use
of the Autocrypt setup message.

We need to handle PLAINTEXT status lines that include a possible
filename in them.  And sometimes GnuPG gets confused about the
different sizes of plaintext: https://dev.gnupg.org/T4741.  so just
warn, don't throw an exception for now as a workaround.