create simple profile during sqlite autopkgtest
I found that the first run of a thunderbird 60 profile with jsunit
installed spews messages to stderr.  The second run appears to work
fine.

I don't understand why that happens, but i'll take advantage of it by
ignoring stderr messages on a simple, minimal, profile-creating
first-run.

Then we'll run the real test.