drop mozilla-devscripts due to lack of install.rdf
The main things we were getting from mozilla-devscripts, afaict, was:

 - extraction of dependency information from install.rdf
 - creation of filesystem uuid links that point to the unpacked .xpi

but enigmail 2.1 no longer ships an install.rdf -- it needs dependency
information extracted from manifest.json.

And the single filesystem uuid link can be handled manually -- this is
only relevant for icedove and thunderbird, which share an app UUID,
afaicr.

So i'm "going solo" with the packaging here, as mozilla-devscripts
doesn't seem to have kept up with the (admittedly unusual) needs of
enigmail.